Utilises the Mininet Python API to build a small test network.
Creates a topology with two hosts, an OpenFlow controller and two switches.

One switch is intended to be used for the actual network between hosts 1 and 2, and the other switch for a management network to remotely control each node. The next step is to ensure that these switches can perform these roles.